requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
filelock==3.13.3
Flask==3.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.11.2
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.14.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
optuna==3.6.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
proto-plus==1.23.0
protobuf==4.25.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.29
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tiktoken==0.6.0
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
tf-keras
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
